OT night =((

+++ SmartChoiceApp/SmartChoiceApp/SmartChoiceApp/Converter/ToDouble.cs
@@ -8,7 +8,7 @@ namespace SmartChoiceApp.Converter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (int.Parse(value.ToString()) *2);
+            return (double.Parse(value.ToString()) *2);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
